pull in preprocessing logic for schedules
