+++ src/app/containers/RadioSchedule/Canonical/index.jsx
@@ -2,6 +2,10 @@ import React, { useEffect, useState, useContext } from 'react';
 import 'isomorphic-fetch';
 import { string } from 'prop-types';
 import styled from 'styled-components';
+import RadioSchedule from '@bbc/psammead-radio-schedule';
+import moment from 'moment';
+import findLastIndex from 'ramda/src/findLastIndex';
+import propSatisfies from 'ramda/src/propSatisfies';
 import SectionLabel from '@bbc/psammead-section-label';
 import { Link } from '@bbc/psammead-story-promo';
 import { ServiceContext } from '#contexts/ServiceContext';
@@ -15,38 +19,100 @@ const RadioFrequencyLink = styled(Link)`
 `;
 
 const CanonicalRadioSchedule = ({ endpoint }) => {
-  const header = 'BBC News Radio';
-  const [schedule, setRadioSchedule] = useState([]);
+  const [schedule, setRadioSchedule] = useState();
   const {
     service,
     script,
     dir,
-    radioSchedule: { frequenciesPageUrl },
+    timezone,
+    locale,
+    radioSchedule: { header, frequenciesPageUrl, frequenciesPageLabel },
   } = useContext(ServiceContext);
 
-  const handleResponse = async response => {
-    const radioScheduleData = await response.json();
-    setRadioSchedule(radioScheduleData.schedules.slice(0, 4));
+  const getProgramState = (currentTime, startTime, endTime) => {
+    const isLive = currentTime < endTime && currentTime > startTime;
+    if (isLive) {
+      return 'live';
+    }
+    const hasEnded = currentTime > endTime;
+    if (hasEnded) {
+      return 'onDemand';
+    }
+    return 'next';
   };
 
   useEffect(() => {
+    const getLink = (state, program) => {
+      const url = `/${service}/${program.serviceId}`;
+      return state === 'live'
+        ? `${url}/liveradio`
+        : `${url}/${program.broadcast.pid}`;
+    };
+
+    const handleResponse = async response => {
+      const radioScheduleData = await response.json();
+
+      const currentTime = parseInt(moment.utc().format('x'), 10);
+      // finding latest program, that may or may not still be live. this is because there isn't
+      // always a live program, in which case we show the most recently played program on demand.
+      const latestProgramIndex = findLastIndex(
+        propSatisfies(time => time < currentTime, 'publishedTimeStart'),
+      )(radioScheduleData.schedules);
+
+      const radioSchedules = radioScheduleData.schedules;
+
+      const scheduleDataIsComplete =
+        radioSchedules[latestProgramIndex - 2] &&
+        radioSchedules[latestProgramIndex + 1];
+      const schedulesToShow = scheduleDataIsComplete && [
+        radioSchedules[latestProgramIndex],
+        radioSchedules[latestProgramIndex - 1],
+        radioSchedules[latestProgramIndex - 2],
+        radioSchedules[latestProgramIndex + 1],
+      ];
+
+      const schedules =
+        schedulesToShow &&
+        schedulesToShow.map(program => {
+          const currentState = getProgramState(
+            currentTime,
+            program.publishedTimeStart,
+            program.publishedTimeEnd,
+          );
+          return {
+            id: program.broadcast.pid,
+            state: currentState,
+            stateLabel: currentState,
+            startTime: program.publishedTimeStart,
+            link: getLink(currentState, program),
+            brandTitle: program.brand.title,
+            episodeTitle: program.episode.presentationTitle,
+            summary: program.episode.synopses.short,
+            duration: program.publishedTimeDuration,
+            durationLabel: 'Duration',
+          };
+        });
+      setRadioSchedule(schedules);
+    };
+
     const fetchRadioScheduleData = pathname =>
       fetch(pathname, { mode: 'no-cors' })
         .then(handleResponse)
         .catch(e => logger.error(`HTTP Error: "${e}"`));
 
     fetchRadioScheduleData(endpoint);
-  }, [endpoint]);
+  }, [endpoint, service, script, timezone, locale]);
 
-  const renderFrequencyLink = link => (
-    <RadioFrequencyLink link={link}>
-      Radio frequency link text
-    </RadioFrequencyLink>
-  );
+  console.log(schedule);
 
-  if (!schedule.length) {
+  if (!schedule) {
     return null;
   }
+
+  const renderFrequencyLink = (link, label) => (
+    <RadioFrequencyLink link={link}>{label}</RadioFrequencyLink>
+  );
+
   return (
     <>
       <SectionLabel
@@ -57,25 +123,18 @@ const CanonicalRadioSchedule = ({ endpoint }) => {
       >
         {header}
       </SectionLabel>
-      {schedule.map(
-        ({
-          broadcast,
-          transmissionTimeStart,
-          transmissionTimeEnd,
-          episode: {
-            presentationTitle,
-            synopses: { short },
-          },
-        }) => (
-          <ul key={broadcast.pid}>
-            <li>{presentationTitle}</li>
-            <li>{short}</li>
-            <li>{transmissionTimeStart}</li>
-            <li>{transmissionTimeEnd}</li>
-          </ul>
-        ),
+      {schedule && (
+        <RadioSchedule
+          schedules={schedule}
+          locale={locale}
+          timezone={timezone}
+          script={script}
+          service={service}
+          dir={dir}
+        />
       )}
-      {frequenciesPageUrl && renderFrequencyLink(frequenciesPageUrl)}
+      {frequenciesPageUrl &&
+        renderFrequencyLink(frequenciesPageUrl, frequenciesPageLabel)}
     </>
   );
 };
